@@ -51,8 +52,15 @@ def jarName = 'rd-caseworker-ref-api.jar'
 group = 'uk.gov.hmcts.reform.rdcaseworker'
 version = '0.0.1'
 
-sourceCompatibility = 11
-targetCompatibility = 11
+java {
+  toolchain {
+    languageVersion = JavaLanguageVersion.of(17)
+  }
+}
+
+tasks.withType(JavaExec).configureEach {
+  javaLauncher.set(javaToolchains.launcherFor(java.toolchain))
+}
 
 jacocoTestCoverageVerification {
   violationRules {
