@@ -12,12 +12,12 @@ import lombok.Setter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import uk.gov.hmcts.reform.lib.config.TestConfig;
 
 @Getter
 @Setter
 @Configuration
-public class TestConfigProperties {
-
+public class TestConfigProperties implements TestConfig {
 
     @Value("${oauth2.client.secret}")
     public String clientSecret;
@@ -45,6 +45,9 @@ public class TestConfigProperties {
 
     @Value("${s2s-secret}")
     protected String s2sSecret;
+
+    @Value("${scope-name}")
+    protected String scope;
 
     @Bean
     public ObjectMapper defaultObjectMapper() {
