@@ -148,9 +148,11 @@ public class CreateCaseWorkerProfilesIntegrationTest extends AuthorizationEnable
         CaseWorkerRoleRequest cwRoleRequest = new CaseWorkerRoleRequest("Regional Centre Administrator", true);
         CaseWorkerRoleRequest cwRoleRequest1 = new CaseWorkerRoleRequest("Regional Centre Team Leader", false);
         CaseWorkerRoleRequest cwRoleRequest2 = new CaseWorkerRoleRequest("DWP Caseworker", false);
+        CaseWorkerRoleRequest cwRoleRequest3 = new CaseWorkerRoleRequest("Registrar", false);
+
 
         List<CaseWorkerRoleRequest> caseWorkerRoleRequests = ImmutableList
-            .of(cwRoleRequest,cwRoleRequest1,cwRoleRequest2);
+            .of(cwRoleRequest,cwRoleRequest1,cwRoleRequest2,cwRoleRequest3);
         caseWorkersProfileCreationRequests.get(0).setRoles(caseWorkerRoleRequests);
         caseWorkersProfileCreationRequests.get(0).setUserType("Other Government Department");
 
@@ -158,14 +160,10 @@ public class CreateCaseWorkerProfilesIntegrationTest extends AuthorizationEnable
             .createCaseWorkerProfile(caseWorkersProfileCreationRequests, "cwd-admin");
         assertThat(response).containsEntry("http_status", "201 CREATED");
         List<CaseWorkerRole> caseWorkerRoles = caseWorkerRoleRepository.findAll();
-        var caseWorkerProfile = caseWorkerProfileRepository.findAll();
         assertEquals(13, (long) caseWorkerRoles.get(0).getRoleId());
         assertEquals(14,(long)caseWorkerRoles.get(2).getRoleId());
+        assertEquals(16,(long)caseWorkerRoles.get(3).getRoleId());
+        var caseWorkerProfile = caseWorkerProfileRepository.findAll();
         assertEquals(5,caseWorkerProfile.get(0).getUserTypeId());
-
-
-
     }
-
-
 }
