@@ -5,10 +5,12 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.annotation.Order;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.oauth2.client.web.OAuth2AuthorizationRequestRedirectFilter;
 import org.springframework.security.oauth2.core.DelegatingOAuth2TokenValidator;
 import org.springframework.security.oauth2.core.OAuth2TokenValidator;
 import org.springframework.security.oauth2.jwt.Jwt;
@@ -40,8 +42,10 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
 
     @Value("${oidc.issuer}")
     private String issuerOverride;
-
+    @Order(1)
     private  ServiceAuthFilter serviceAuthFilter;
+    @Order(2)
+    private final SecurityEndpointFilter securityEndpointFilter;
     List<String> anonymousPaths;
 
     private JwtAuthenticationConverter jwtAuthenticationConverter;
@@ -65,12 +69,14 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
     @Inject
     public SecurityConfiguration(final JwtGrantedAuthoritiesConverter jwtGrantedAuthoritiesConverter,
                                  final ServiceAuthFilter serviceAuthFilter,
-                                 RestAuthenticationEntryPoint restAuthenticationEntryPoint) {
+                                 RestAuthenticationEntryPoint restAuthenticationEntryPoint,
+                                 SecurityEndpointFilter securityEndpointFilter) {
 
         this.serviceAuthFilter = serviceAuthFilter;
         this.restAuthenticationEntryPoint = restAuthenticationEntryPoint;
         jwtAuthenticationConverter = new JwtAuthenticationConverter();
         jwtAuthenticationConverter.setJwtGrantedAuthoritiesConverter(jwtGrantedAuthoritiesConverter);
+        this.securityEndpointFilter = securityEndpointFilter;
 
     }
 
@@ -78,6 +84,7 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
     protected void configure(final HttpSecurity http) throws Exception {
         http
            .addFilterBefore(serviceAuthFilter, BearerTokenAuthenticationFilter.class)
+           .addFilterAfter(securityEndpointFilter, OAuth2AuthorizationRequestRedirectFilter.class)
            .sessionManagement().sessionCreationPolicy(STATELESS).and()
            .csrf().disable()
            .formLogin().disable()
