@@ -20,14 +20,14 @@ class FeatureToggleServiceImplTest {
     }
 
     @Test
-    public void testIsFlagEnabled_true() {
+    void testIsFlagEnabled_true() {
         flaFeatureToggleService = new FeatureToggleServiceImpl(ldClient, "rd");
         when(flaFeatureToggleService.isFlagEnabled("test", "test")).thenReturn(true);
         assertTrue(flaFeatureToggleService.isFlagEnabled("test", "test"));
     }
 
     @Test
-    public void mapServiceToFlagTest() {
+    void mapServiceToFlagTest() {
         flaFeatureToggleService = new FeatureToggleServiceImpl(ldClient, "rd");
         flaFeatureToggleService.mapServiceToFlag();
         assertTrue(flaFeatureToggleService.getLaunchDarklyMap().size() >= 1);
